Integrate loading buttons in auth components, add form directives to login component
