+++ client/src/app/core/change-password/change-password.component.html
@@ -39,7 +39,7 @@
                *ngIf="changePasswordPressed && passwordsDontMatch()">
     </p-message>
 
-    <p-button label="Passwort ändern" (click)="changePassword()"></p-button>
+    <p-button label="Passwort ändern" (click)="changePassword()" [loading]="loading"></p-button>
 
   </div>
