+++ server/tests/test_auth_resources.py
@@ -221,7 +221,7 @@ def test_token_user_does_not_exist(client):
         'oldPassword': '[vb+xLGgU?+Z]nXD3HmO',
         'newPassword': 'testPassword'
     }
-    rv = client.put('/api/account/change-password', headers=access_headers, json=change_pw_data)
+    rv = client.put('/api/change-password', headers=access_headers, json=change_pw_data)
     assert rv.status_code == 401
     res = json.loads(rv.data)
     assert res['message'] == ResponseMessage.UNAUTHORIZED.value
@@ -237,7 +237,7 @@ def test_revoked_access_token_bahaviour(client):
         'oldPassword': 'kevin123!',
         'newPassword': 'testPassword'
     }
-    rv = client.put('/api/account/change-password', headers=access_headers, json=change_pw_data)
+    rv = client.put('/api/change-password', headers=access_headers, json=change_pw_data)
     assert rv.status_code == 401
 
 
@@ -303,3 +303,40 @@ def test_reset_password_with_non_activated_user(client):
     assert rv.status_code == 401
     res = json.loads(rv.data)
     assert res['message'] == ResponseMessage.USER_NOT_ACTIVATED.value
+
+
+
+def test_successful_change_password(client):
+    access_headers, refresh_headers = get_login_headers(client)
+    change_pw_data = {
+        'oldPassword': '[vb+xLGgU?+Z]nXD3HmO',
+        'newPassword': 'fengelmann2'
+    }
+    rv = client.put('/api/change-password', headers=access_headers, json=change_pw_data)
+    assert rv.status_code == 201
+    res = json.loads(rv.data)
+    assert res['message'] == ResponseMessage.PASSWORD_CHANGED.value
+
+
+def test_change_password_password_too_short(client):
+    access_headers, refresh_headers = get_login_headers(client)
+    change_pw_data = {
+        'oldPassword': '[vb+xLGgU?+Z]nXD3HmO',
+        'newPassword': '1234567'
+    }
+    rv = client.put('/api/change-password', headers=access_headers, json=change_pw_data)
+    assert rv.status_code == 400
+    res = json.loads(rv.data)
+    assert res['message'] == ResponseMessage.PASSWORD_TOO_SHORT.value
+
+
+def test_change_password_password_old_pw_incorrect(client):
+    access_headers, refresh_headers = get_login_headers(client)
+    change_pw_data = {
+        'oldPassword': 'incorrectpassword',
+        'newPassword': 'fengelmann2'
+    }
+    rv = client.put('/api/change-password', headers=access_headers, json=change_pw_data)
+    assert rv.status_code == 401
+    res = json.loads(rv.data)
+    assert res['message'] == ResponseMessage.OLD_PASSWORD_INCORRECT.value
